Create-edit: added if statement to show delete button only in edit post

+++ app/views/lostItems/create-edit.blade.php
@@ -2,7 +2,7 @@
 
 @section('content')
 
-@if (empty($lostItem->id))
+@if (empty($lostItems->id))
 
 <div class="row">
     <div class="col-sm-12">
@@ -20,7 +20,7 @@
 	</div>
 </div>
 
-	{{ Form::model($lostItem, array('action' => array('LostItemsController@update', $lostItem->id), 'method' => 'PUT', 'files' => true, 'class' => 'form-horizontal')) }}
+	{{ Form::model($lostItems, array('action' => array('LostItemsController@update', $lostItems->id), 'method' => 'PUT', 'files' => true, 'class' => 'form-horizontal')) }}
 
 @endif
 		<div class="form-group">
@@ -78,11 +78,13 @@
 		   		<a href="{{{action('LostItemsController@index') }}}" style='text-decoration:none;color:#FFF'><button class='btn btn-primary'>Cancel</a>
  				@endif	   
 		    </div>
-		    <div class="col-sm-offset-2 col-sm-10">
-		    	{{ Form::open(array('action' => array('LostItemsController@destroy', $lostItems->id), 'method' => 'delete')) }}
-				{{ Form::submit('Delete', array('class' => 'btnDelete btn btn-danger')) }}
-				{{ Form::close() }}
-			</div>
+		    @if (!empty($foundItems->id))
+		    	<div class="col-sm-offset-2 col-sm-10">
+		    		{{ Form::open(array('action' => array('LostItemsController@destroy', $lostItems->id), 'method' => 'delete')) }}
+					{{ Form::submit('Delete', array('class' => 'btnDelete btn btn-danger')) }}
+					{{ Form::close() }}
+				</div>
+			@endif
 		</div>
 
 @stop
